requirements.txt
ecdsa>=0.18.0
pyobjc-core>=10.3
pyobjc-framework-Metal>=10.3

